# packages/server/openapi/v2.yaml
...
openapi: 3.1.1
info:
  title: Memax API v2
  version: 2.0.0-alpha.1
  summary: The record your agents read from, with a receipt for every change.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The `/v2` API reads and changes the V2 record: memories, their versions,
    sources and receipts, the Brief, and the targets it compiles to, in the
    spaces you belong to.

    ## Envelope

    Every response is JSON. A success carries `{"data": …}`. An error carries
    `{"error": {"code", "message", "details"}}`. `code` is stable and
    machine-readable; `message` is an English sentence that says what to do.

    ## Commands

    A command is a `POST` (or the `PATCH` that sets an agent's autonomy or
    changes a target) that changes the record: remember, keep, edit and
    reject a memory; pause, resume, disconnect and set the autonomy of an
    agent; revise the Brief; add, change and compile a target; acknowledge
    a delivery, report a hand edit, and pull, overwrite or stop it. Every
    command:

    - needs an `Idempotency-Key` header (1 to 255 characters). Retrying a
      command with the same key returns the original result and writes
      nothing; the response then carries `Idempotent-Replayed: true`. A key
      reused for a different command or different content is refused with
      422 `idempotency_key_reused`.
    - writes one or more receipts in the same transaction. A receipt never
      holds the memory's words.
...

...

...

...
    seen.

...

...

...
servers:
...
security:
  - bearer: []
...
paths:
...
  /v2/spaces/{space}/brief:
    parameters:
      - $ref: '#/components/parameters/SpacePath'
    get:
      operationId: getBrief
      tags: [briefs]
      summary: Read the Brief
      description: |
        The space's current Brief version, with the receipt of whoever wrote
        it. A memory item is a display ID; render its statement from the
        memory itself. 404 when the space has no Brief yet.
      responses:
        '200':
          description: The current version.
          headers:
            ETag:
              $ref: '#/components/headers/VersionETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: reviseBrief
      tags: [briefs]
      summary: Revise the Brief
      description: |
        Writes a new version (`B-`) of the space's Brief: ordered sections of
        memory items and cited prose. Every memory item must be a kept
        memory of the space; every prose line must cite at least one kept
        memory, and nothing forgotten or rejected. Send `If-Match` with the
        version you started from; it is required once the space has a
        Brief. People who may keep revise the Brief; agents can't. Every
        target of the space recompiles.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
        - $ref: '#/components/parameters/Via'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseBriefRequest'
      responses:
        '201':
          description: The new version.
          headers:
            ETag:
              $ref: '#/components/headers/VersionETag'
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefResultEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/EditClash'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/spaces/{space}/brief/versions:
    parameters:
      - $ref: '#/components/parameters/SpacePath'
    get:
      operationId: listBriefVersions
      tags: [briefs]
      summary: List the Brief's versions
      description: |
        Every version of the space's Brief, newest first, each with the
        receipt of whoever wrote it and why (BriefHistory). Restore one by
        revising the Brief with its sections.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: One page of versions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefVersionPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/spaces/{space}/targets:
    parameters:
      - $ref: '#/components/parameters/SpacePath'
    get:
      operationId: listTargets
      tags: [targets]
      summary: List where the space compiles to
      description: |
        The space's targets (AGENTS.md, the CLAUDE.md shim, scoped Cursor
        rules, the ChatGPT copy-out, …), each with its sync state, its latest
        compile run and what is on disk. Not paginated.
      responses:
        '200':
          description: The targets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createTarget
      tags: [targets]
      summary: Add a target
      description: |
        Adds a compile target to the space, with the kind's defaults for
        whatever you leave out (path, settings, delivery), and compiles it.
        The space needs a Brief first. People who may keep add targets.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Via'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '201':
          description: The target, compiling.
          headers:
            ETag:
              $ref: '#/components/headers/VersionETag'
            Location:
              $ref: '#/components/headers/TargetLocation'
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResultEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/targets/{target}:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    patch:
      operationId: configureTarget
      tags: [targets]
      summary: Change a target
      description: |
        Changes a target's path, settings or delivery, or stops and restarts
        compiling it (`enabled`), and recompiles it. Send `If-Match` with
        the target's version to be sure you change what you saw. A change
        that changes nothing writes nothing and has no receipts.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
        - $ref: '#/components/parameters/Via'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigureTargetRequest'
      responses:
        '200':
          $ref: '#/components/responses/TargetResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/EditClash'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/targets/{target}:compile:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    post:
      operationId: compileTarget
      tags: [targets]
      summary: Compile a target now
      description: |
        Asks for a fresh compile of the target (Compile now). It is queued,
        so the answer is 202 with the target compiling; follow
        `last_compile` on the target, or the runs. A stopped target is 409.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Via'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '202':
          $ref: '#/components/responses/TargetResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InvalidTransition'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/targets/{target}/preview:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    get:
      operationId: getTargetPreview
      tags: [targets]
      summary: Read what a target compiled to
      description: |
        The latest good compile's content (TargetPreview): every file it
        writes, or the copy-out text, with the run's metadata. Before the
        first compile, `compile` is absent and there are no files. 503 when
        the server has no compile service.
      responses:
        '200':
          description: The compiled content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetPreviewEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/targets/{target}/runs:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    get:
      operationId: listCompileRuns
      tags: [targets]
      summary: List a target's compile runs
      description: Every compile run (`C-`) of the target, newest first, failed ones included.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: One page of runs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileRunPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v2/targets/{target}/observations:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    post:
      operationId: recordObservation
      tags: [targets]
      summary: Report a target's file as seen
      description: |
        A device (the local daemon) reports a file the target writes, as it
        is on disk. When it matches what Memax delivered (line endings and a
        BOM aside), nothing is recorded: 200 with `drifted: false`. Otherwise
        it is a hand edit: the content is kept, the compiler reads it back
        as changes, and the target is drifted until a person pulls,
        overwrites or stops it: 201 with the observation.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Via'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationRequest'
      responses:
        '200':
          description: The file matches what Memax delivered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResultEnvelope'
        '201':
          description: A hand edit, recorded.
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResultEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /v2/targets/{target}/drift:
    parameters:
      - $ref: '#/components/parameters/TargetPath'
    get:
      operationId: getDrift
      tags: [targets]
      summary: Read a target's hand edits
      description: |
        Each open hand edit of the target (DriftResolve): the file as Memax
        wrote it, the file as it is now, and what the edit says, change by
        change. No open edits is an empty list. 503 when the server has no
        compile service.
      responses:
        '200':
          description: The open hand edits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        A signed-in session token, an OAuth access token, or an API key
        (`mxk_…`). A session acts as the person. An API key or an OAuth
        grant acts as the agent connection it is bound to, at that
        connection's autonomy in each space (Read when the credential has no
        write access; an API key proposes at most). A credential with no
        connection, or a paused one, only reads.
  parameters:
    SpacePath:
      name: space
      in: path
      required: true
      description: The space's id or slug.
      schema:
        $ref: '#/components/schemas/SpaceKey'
...
  headers:
...
    VersionETag:
      description: The object's version, quoted. Send it back as `If-Match`.
      required: true
      schema:
        $ref: '#/components/schemas/VersionTag'
...
  responses:
...
    BadRequest:
      description: |
        `invalid_request`, `idempotency_key_required`, `space_required` or
        `ambiguous_ref`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: '`unauthorized`: sign in again or check the API key.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    SpaceKey:
      description: A space's id or slug.
      type: string
      minLength: 1
      maxLength: 255
...
    BriefEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Brief'
...